EMP/my_work/ad_login/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
python-multipart
mysqlclient
mysql-connector-python
mysql